test(app): migrate safe-area
